Update variable types in soal1.c to double
Changed variable types from int to double for precision.

+++ soal1.c
@@ -1,6 +1,6 @@
 int main() {
-    int a, b, c;
-    int max;
+    double a, b, c;
+    double max;
 
     printf("Masukkan sisi segitiga: ");
     scanf("%d %d %d", &a, &b, &c);
